refactor conformance-tests binary
