+++ cmd/conformance-tester/pkg/scenarios/alibaba.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2020 The Kubermatic Kubernetes Platform contributors.
+Copyright 2022 The Kubermatic Kubernetes Platform contributors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,29 +14,32 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package scenarios
 
 import (
 	"context"
 	"fmt"
 
+	"k8c.io/kubermatic/v2/cmd/conformance-tester/pkg/types"
 	"k8c.io/kubermatic/v2/pkg/semver"
 	apimodels "k8c.io/kubermatic/v2/pkg/test/e2e/utils/apiclient/models"
 )
 
-// Returns a matrix of (version x operating system).
-func getAzureScenarios(versions []*semver.Semver) []testScenario {
-	var scenarios []testScenario
+const alibabaDC = "alibaba-eu-central-1a"
+
+// GetAlibabaScenarios returns a matrix of (version x operating system).
+func GetAlibabaScenarios(versions []*semver.Semver) []Scenario {
+	var scenarios []Scenario
 	for _, v := range versions {
 		// Ubuntu
-		scenarios = append(scenarios, &azureScenario{
+		scenarios = append(scenarios, &alibabaScenario{
 			version: v,
 			nodeOsSpec: apimodels.OperatingSystemSpec{
 				Ubuntu: &apimodels.UbuntuSpec{},
 			},
 		})
 		// CentOS
-		scenarios = append(scenarios, &azureScenario{
+		scenarios = append(scenarios, &alibabaScenario{
 			version: v,
 			nodeOsSpec: apimodels.OperatingSystemSpec{
 				Centos: &apimodels.CentOSSpec{},
@@ -46,27 +49,25 @@ func getAzureScenarios(versions []*semver.Semver) []testScenario {
 	return scenarios
 }
 
-type azureScenario struct {
+type alibabaScenario struct {
 	version    *semver.Semver
 	nodeOsSpec apimodels.OperatingSystemSpec
 }
 
-func (s *azureScenario) Name() string {
-	return fmt.Sprintf("azure-%s-%s", getOSNameFromSpec(s.nodeOsSpec), s.version.String())
+func (s *alibabaScenario) Name() string {
+	return fmt.Sprintf("alibaba-%s-%s", getOSNameFromSpec(s.nodeOsSpec), s.version.String())
 }
 
-func (s *azureScenario) Cluster(secrets secrets) *apimodels.CreateClusterSpec {
+func (s *alibabaScenario) Cluster(secrets types.Secrets) *apimodels.CreateClusterSpec {
 	return &apimodels.CreateClusterSpec{
 		Cluster: &apimodels.Cluster{
 			Type: "kubernetes",
 			Spec: &apimodels.ClusterSpec{
 				Cloud: &apimodels.CloudSpec{
-					DatacenterName: "azure-westeurope",
-					Azure: &apimodels.AzureCloudSpec{
-						ClientID:       secrets.Azure.ClientID,
-						ClientSecret:   secrets.Azure.ClientSecret,
-						SubscriptionID: secrets.Azure.SubscriptionID,
-						TenantID:       secrets.Azure.TenantID,
+					DatacenterName: alibabaDC,
+					Alibaba: &apimodels.AlibabaCloudSpec{
+						AccessKeySecret: secrets.Alibaba.AccessKeySecret,
+						AccessKeyID:     secrets.Alibaba.AccessKeyID,
 					},
 				},
 				Version: apimodels.Semver(s.version.String()),
@@ -75,17 +76,19 @@ func (s *azureScenario) Cluster(secrets secrets) *apimodels.CreateClusterSpec {
 	}
 }
 
-func (s *azureScenario) NodeDeployments(_ context.Context, num int, _ secrets) ([]apimodels.NodeDeployment, error) {
-	replicas := int32(num)
-	size := "Standard_F2"
+func (s *alibabaScenario) NodeDeployments(_ context.Context, num int, secrets types.Secrets) ([]apimodels.NodeDeployment, error) {
 	return []apimodels.NodeDeployment{
 		{
 			Spec: &apimodels.NodeDeploymentSpec{
-				Replicas: &replicas,
 				Template: &apimodels.NodeSpec{
 					Cloud: &apimodels.NodeCloudSpec{
-						Azure: &apimodels.AzureNodeSpec{
-							Size: &size,
+						Alibaba: &apimodels.AlibabaNodeSpec{
+							InstanceType:            "ecs.c6.xsmall",
+							DiskSize:                "40",
+							DiskType:                "cloud_efficiency",
+							VSwitchID:               "vsw-gw8g8mn4ohmj483hsylmn",
+							InternetMaxBandwidthOut: "10",
+							ZoneID:                  alibabaDC,
 						},
 					},
 					Versions: &apimodels.NodeVersionInfo{
@@ -98,6 +101,6 @@ func (s *azureScenario) NodeDeployments(_ context.Context, num int, _ secrets) (
 	}, nil
 }
 
-func (s *azureScenario) OS() apimodels.OperatingSystemSpec {
+func (s *alibabaScenario) OS() apimodels.OperatingSystemSpec {
 	return s.nodeOsSpec
 }
