+++ cmd/conformance-tester/pkg/scenarios/azure.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2020 The Kubermatic Kubernetes Platform contributors.
+Copyright 2022 The Kubermatic Kubernetes Platform contributors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,58 +14,60 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package scenarios
 
 import (
 	"context"
 	"fmt"
 
+	"k8c.io/kubermatic/v2/cmd/conformance-tester/pkg/types"
 	"k8c.io/kubermatic/v2/pkg/semver"
 	apimodels "k8c.io/kubermatic/v2/pkg/test/e2e/utils/apiclient/models"
 )
 
-// Returns a matrix of (version x operating system).
-func getPacketScenarios(versions []*semver.Semver) []testScenario {
-	var scenarios []testScenario
+// GetAzureScenarios returns a matrix of (version x operating system).
+func GetAzureScenarios(versions []*semver.Semver) []Scenario {
+	var scenarios []Scenario
 	for _, v := range versions {
 		// Ubuntu
-		scenarios = append(scenarios, &packetScenario{
+		scenarios = append(scenarios, &azureScenario{
 			version: v,
 			nodeOsSpec: apimodels.OperatingSystemSpec{
 				Ubuntu: &apimodels.UbuntuSpec{},
 			},
 		})
 		// CentOS
-		scenarios = append(scenarios, &packetScenario{
+		scenarios = append(scenarios, &azureScenario{
 			version: v,
 			nodeOsSpec: apimodels.OperatingSystemSpec{
 				Centos: &apimodels.CentOSSpec{},
 			},
 		})
 	}
-
 	return scenarios
 }
 
-type packetScenario struct {
+type azureScenario struct {
 	version    *semver.Semver
 	nodeOsSpec apimodels.OperatingSystemSpec
 }
 
-func (s *packetScenario) Name() string {
-	return fmt.Sprintf("packet-%s-%s", getOSNameFromSpec(s.nodeOsSpec), s.version.String())
+func (s *azureScenario) Name() string {
+	return fmt.Sprintf("azure-%s-%s", getOSNameFromSpec(s.nodeOsSpec), s.version.String())
 }
 
-func (s *packetScenario) Cluster(secrets secrets) *apimodels.CreateClusterSpec {
+func (s *azureScenario) Cluster(secrets types.Secrets) *apimodels.CreateClusterSpec {
 	return &apimodels.CreateClusterSpec{
 		Cluster: &apimodels.Cluster{
 			Type: "kubernetes",
 			Spec: &apimodels.ClusterSpec{
 				Cloud: &apimodels.CloudSpec{
-					DatacenterName: "packet-ewr1",
-					Packet: &apimodels.PacketCloudSpec{
-						APIKey:    secrets.Packet.APIKey,
-						ProjectID: secrets.Packet.ProjectID,
+					DatacenterName: "azure-westeurope",
+					Azure: &apimodels.AzureCloudSpec{
+						ClientID:       secrets.Azure.ClientID,
+						ClientSecret:   secrets.Azure.ClientSecret,
+						SubscriptionID: secrets.Azure.SubscriptionID,
+						TenantID:       secrets.Azure.TenantID,
 					},
 				},
 				Version: apimodels.Semver(s.version.String()),
@@ -74,17 +76,17 @@ func (s *packetScenario) Cluster(secrets secrets) *apimodels.CreateClusterSpec {
 	}
 }
 
-func (s *packetScenario) NodeDeployments(_ context.Context, num int, _ secrets) ([]apimodels.NodeDeployment, error) {
-	instanceType := "t1.small.x86"
+func (s *azureScenario) NodeDeployments(_ context.Context, num int, _ types.Secrets) ([]apimodels.NodeDeployment, error) {
 	replicas := int32(num)
+	size := "Standard_F2"
 	return []apimodels.NodeDeployment{
 		{
 			Spec: &apimodels.NodeDeploymentSpec{
 				Replicas: &replicas,
 				Template: &apimodels.NodeSpec{
 					Cloud: &apimodels.NodeCloudSpec{
-						Packet: &apimodels.PacketNodeSpec{
-							InstanceType: &instanceType,
+						Azure: &apimodels.AzureNodeSpec{
+							Size: &size,
 						},
 					},
 					Versions: &apimodels.NodeVersionInfo{
@@ -97,6 +99,6 @@ func (s *packetScenario) NodeDeployments(_ context.Context, num int, _ secrets) 
 	}, nil
 }
 
-func (s *packetScenario) OS() apimodels.OperatingSystemSpec {
+func (s *azureScenario) OS() apimodels.OperatingSystemSpec {
 	return s.nodeOsSpec
 }
