+++ cmd/conformance-tester/pkg/scenarios/kubevirt.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2020 The Kubermatic Kubernetes Platform contributors.
+Copyright 2022 The Kubermatic Kubernetes Platform contributors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package scenarios
 
 import (
 	"context"
@@ -22,15 +22,16 @@ import (
 
 	"go.uber.org/zap"
 
+	"k8c.io/kubermatic/v2/cmd/conformance-tester/pkg/types"
 	"k8c.io/kubermatic/v2/pkg/semver"
 	apimodels "k8c.io/kubermatic/v2/pkg/test/e2e/utils/apiclient/models"
 
 	utilpointer "k8s.io/utils/pointer"
 )
 
-// Returns a matrix of (version x operating system).
-func getKubevirtScenarios(versions []*semver.Semver, log *zap.SugaredLogger) []testScenario {
-	var scenarios []testScenario
+// GetKubevirtScenarios Returns a matrix of (version x operating system).
+func GetKubevirtScenarios(versions []*semver.Semver, log *zap.SugaredLogger) []Scenario {
+	var scenarios []Scenario
 	for _, v := range versions {
 		// Ubuntu
 		scenarios = append(scenarios, &kubevirtScenario{
@@ -63,7 +64,7 @@ func (s *kubevirtScenario) Name() string {
 	return fmt.Sprintf("kubevirt-%s-%s", getOSNameFromSpec(*s.nodeOsSpec), s.version.String())
 }
 
-func (s *kubevirtScenario) Cluster(secrets secrets) *apimodels.CreateClusterSpec {
+func (s *kubevirtScenario) Cluster(secrets types.Secrets) *apimodels.CreateClusterSpec {
 	return &apimodels.CreateClusterSpec{
 		Cluster: &apimodels.Cluster{
 			Type: "kubernetes",
@@ -80,7 +81,7 @@ func (s *kubevirtScenario) Cluster(secrets secrets) *apimodels.CreateClusterSpec
 	}
 }
 
-func (s *kubevirtScenario) NodeDeployments(_ context.Context, num int, _ secrets) ([]apimodels.NodeDeployment, error) {
+func (s *kubevirtScenario) NodeDeployments(_ context.Context, num int, _ types.Secrets) ([]apimodels.NodeDeployment, error) {
 	var sourceURL string
 	registryAddr := "http://10.244.2.19"
