+++ cmd/conformance-tester/pkg/scenarios/packet.go
@@ -1,5 +1,5 @@
 /*
-Copyright 2020 The Kubermatic Kubernetes Platform contributors.
+Copyright 2022 The Kubermatic Kubernetes Platform contributors.
 
 Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
@@ -14,60 +14,59 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package main
+package scenarios
 
 import (
 	"context"
 	"fmt"
-	"strings"
 
+	"k8c.io/kubermatic/v2/cmd/conformance-tester/pkg/types"
 	"k8c.io/kubermatic/v2/pkg/semver"
 	apimodels "k8c.io/kubermatic/v2/pkg/test/e2e/utils/apiclient/models"
 )
 
-// Returns a matrix of (version x operating system).
-func getGCPScenarios(versions []*semver.Semver) []testScenario {
-	var scenarios []testScenario
+// GetPacketScenarios returns a matrix of (version x operating system).
+func GetPacketScenarios(versions []*semver.Semver) []Scenario {
+	var scenarios []Scenario
 	for _, v := range versions {
 		// Ubuntu
-		scenarios = append(scenarios, &gcpScenario{
+		scenarios = append(scenarios, &packetScenario{
 			version: v,
 			nodeOsSpec: apimodels.OperatingSystemSpec{
 				Ubuntu: &apimodels.UbuntuSpec{},
 			},
 		})
 		// CentOS
-		scenarios = append(scenarios, &gcpScenario{
+		scenarios = append(scenarios, &packetScenario{
 			version: v,
 			nodeOsSpec: apimodels.OperatingSystemSpec{
 				Centos: &apimodels.CentOSSpec{},
 			},
 		})
 	}
+
 	return scenarios
 }
 
-type gcpScenario struct {
+type packetScenario struct {
 	version    *semver.Semver
 	nodeOsSpec apimodels.OperatingSystemSpec
 }
 
-func (s *gcpScenario) Name() string {
-	version := strings.ReplaceAll(s.version.String(), ".", "-")
-	return fmt.Sprintf("gcp-%s-%s", getOSNameFromSpec(s.nodeOsSpec), version)
+func (s *packetScenario) Name() string {
+	return fmt.Sprintf("packet-%s-%s", getOSNameFromSpec(s.nodeOsSpec), s.version.String())
 }
 
-func (s *gcpScenario) Cluster(secrets secrets) *apimodels.CreateClusterSpec {
+func (s *packetScenario) Cluster(secrets types.Secrets) *apimodels.CreateClusterSpec {
 	return &apimodels.CreateClusterSpec{
 		Cluster: &apimodels.Cluster{
 			Type: "kubernetes",
 			Spec: &apimodels.ClusterSpec{
 				Cloud: &apimodels.CloudSpec{
-					DatacenterName: "gcp-westeurope",
-					Gcp: &apimodels.GCPCloudSpec{
-						ServiceAccount: secrets.GCP.ServiceAccount,
-						Network:        secrets.GCP.Network,
-						Subnetwork:     secrets.GCP.Subnetwork,
+					DatacenterName: "packet-ewr1",
+					Packet: &apimodels.PacketCloudSpec{
+						APIKey:    secrets.Packet.APIKey,
+						ProjectID: secrets.Packet.ProjectID,
 					},
 				},
 				Version: apimodels.Semver(s.version.String()),
@@ -76,24 +75,17 @@ func (s *gcpScenario) Cluster(secrets secrets) *apimodels.CreateClusterSpec {
 	}
 }
 
-func (s *gcpScenario) NodeDeployments(_ context.Context, num int, secrets secrets) ([]apimodels.NodeDeployment, error) {
+func (s *packetScenario) NodeDeployments(_ context.Context, num int, _ types.Secrets) ([]apimodels.NodeDeployment, error) {
+	instanceType := "t1.small.x86"
 	replicas := int32(num)
-
 	return []apimodels.NodeDeployment{
 		{
 			Spec: &apimodels.NodeDeploymentSpec{
 				Replicas: &replicas,
 				Template: &apimodels.NodeSpec{
 					Cloud: &apimodels.NodeCloudSpec{
-						Gcp: &apimodels.GCPNodeSpec{
-							Zone:        secrets.GCP.Zone,
-							MachineType: "n1-standard-2",
-							DiskType:    "pd-standard",
-							DiskSize:    50,
-							Preemptible: false,
-							Labels: map[string]string{
-								"kubernetes-cluster": "my-cluster",
-							},
+						Packet: &apimodels.PacketNodeSpec{
+							InstanceType: &instanceType,
 						},
 					},
 					Versions: &apimodels.NodeVersionInfo{
@@ -106,6 +98,6 @@ func (s *gcpScenario) NodeDeployments(_ context.Context, num int, secrets secret
 	}, nil
 }
 
-func (s *gcpScenario) OS() apimodels.OperatingSystemSpec {
+func (s *packetScenario) OS() apimodels.OperatingSystemSpec {
 	return s.nodeOsSpec
 }
